Edited SecurityFilterChain method for managing stateless sessions, moved class to new package
